% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biigle_add_label.R
\name{add_unique_column_values_as_labels_bulk}
\alias{add_unique_column_values_as_labels_bulk}
\title{Adds a complete label branch to a BIIGLE label tree. The bulk version addes the children as a single API call}
\usage{
add_unique_column_values_as_labels_bulk(
  biigle_api_connection,
  image_metadata,
  label_tree_id,
  column_name
)
}
\arguments{
\item{biigle_api_connection}{An api connection helper object. Use biigle_api_connect() to create one.}

\item{image_metadata}{The the dataframe having the labels under a columns}

\item{label_tree_id}{The label tree id.}

\item{column_name}{The name of the column in image_metadata that contains the labels to add.}
}
\description{
The Label branch comprises a parent node, and contains all unique values under column_name from image_metadata as child nodes.
This makes a POST API call to api/v1/label-trees/:id/labels
https://biigle.de/doc/api/index.html#api-Label_Trees-StoreLabelTreesLabels
}
\details{
Example, this is usefull to add all unique scientific names as labels under a parent node "scientific_name"
And will allow to tag images with their scientific names from this tree.
}
