% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biigle_add_label.R
\name{biigle_add_bulk_labels}
\alias{biigle_add_bulk_labels}
\title{Add a multiple labels to a BIIGLE label tree}
\usage{
biigle_add_bulk_labels(
  biigle_api_connection,
  label_tree_id,
  parent_label_name,
  label_list,
  color = "#FF0000"
)
}
\arguments{
\item{biigle_api_connection}{An api connection helper object. Use biigle_api_connect() to create one.}

\item{label_tree_id}{The label tree ID.}

\item{parent_label_name}{The label string of the new parent label.}

\item{label_list}{List of label strings to add as children of parent_label. Please make sure to omit the NA values from this list.}

\item{color}{The color code for the labels, default is red "#FF0000"}
}
\description{
This makes a POST API call to api/v1/label-trees/:id/merge-labels
https://biigle.de/doc/api/index.html#api-Label_Trees-StoreLabelTreesMergeLabels
}
