% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biigle_add_label.R
\name{biigle_add_label}
\alias{biigle_add_label}
\title{Add a single label to a BIIGLE label tree}
\usage{
biigle_add_label(
  biigle_api_connection,
  label_tree_id,
  label_name,
  parent_label_id = NULL,
  color = "#FF0000"
)
}
\arguments{
\item{biigle_api_connection}{An api connection helper object. Use biigle_api_connect() to create one.}

\item{label_tree_id}{The label tree ID.}

\item{label_name}{The label string.}

\item{parent_label_id}{The id of the parent label.}

\item{color}{The color code for the labels, default is red "#FF0000"}
}
\description{
This makes a POST API call to api/v1/label-trees/:id/labels
https://biigle.de/doc/api/index.html#api-Label_Trees-StoreLabelTreesLabels
}
