% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biigle_api_connect.R
\name{biigle_api_connect}
\alias{biigle_api_connect}
\title{Generate an object containing authenfication header to use the BIIGLE REST-API}
\usage{
biigle_api_connect(base_url, email, token)
}
\arguments{
\item{base_url}{The base url for the BIIGLE instance.}

\item{email}{The BIIGLE user email.}

\item{token}{The BIIGLE user API token.}
}
\value{
A biigle_api_connection object needed to make subsequent API calls.
}
\description{
This function builds a helper object containing all needed information to make REST-API calls.
It is needed to create every request.
}
