% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biigle_get_images.R
\name{biigle_get_images}
\alias{biigle_get_images}
\title{Gets image metadata as dataframe from a database query to ANDES}
\usage{
biigle_get_images(biigle_api_connection, volume_id)
}
\arguments{
\item{biigle_api_connection}{An api connection helper object. Use biigle_api_connect() to create one.}

\item{volume_id}{The BIIGLE volume id to get images from.}
}
\value{
A dataframe containing the images in the volume
}
\description{
This function executes a REST-API call to a BIIGLE instance to retrieve the images a volume.
A GET request is made to the path api/v1/volumes/:id/filenames
See https://biigle.de/doc/api/index.html#api-Volumes-VolumeIndexFilenames
}
