% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biigle_get_labels.R
\name{biigle_get_labels}
\alias{biigle_get_labels}
\title{Gets labels from a BIIGLE label tree}
\usage{
biigle_get_labels(biigle_api_connection, label_tree_id)
}
\arguments{
\item{biigle_api_connection}{An api connection helper object. Use biigle_api_connect() to create one.}

\item{label_tree_id}{The BIIGLE label_tree id to get the labels from.}
}
\value{
A dataframe containing the labels in the label tree. (with ID and text)
}
\description{
This function executes a REST-API call to a BIIGLE instance to retrieve the labels in a labeltree.
A GET request is made to the path api/v1/label-trees/:id
See https://biigle.de/doc/api/index.html#api-Label_Trees-ShowLabelTrees
}
