% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biigle_label_image.R
\name{biigle_label_image}
\alias{biigle_label_image}
\title{Add a label to a BIIGLE image}
\usage{
biigle_label_image(biigle_api_connection, image_id, label_id)
}
\arguments{
\item{biigle_api_connection}{An api connection helper object. Use biigle_api_connect() to create one.}

\item{image_id}{The id of the image.}

\item{label_id}{The id of the label.}
}
\description{
This function associates a lable (referenced by label_id) to an image (referenced by image_id).
A POST request is made to api/v1/images/:id/labels
See https://biigle.de/doc/api/index.html#api-Images-StoreImageLabels
}
