% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image_metadata.R
\name{get_image_metadata}
\alias{get_image_metadata}
\title{Gets image metadata as dataframe from a database query to ANDES}
\usage{
get_image_metadata(andes_db_connection)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}
}
\value{
A dataframe containing the image metadata: filename, aphia_id, scientific_name, set_number, latitude, longitude, station_name, code.
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct the image metadata table.
The current ANDES active mission will determine for which data are returned.
}
