% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_label_id.R
\name{merge_label_id}
\alias{merge_label_id}
\title{Merge a label_id column}
\usage{
merge_label_id(image_metadata, biigle_labels, column_name)
}
\arguments{
\item{image_metadata}{a dataframe containing a \code{column_name} (and normally an "image_id") column}

\item{biigle_labels}{a dataframe containing the "name" and "label_id" column}

\item{column_name}{The name that the columns containing the labels, a new column having "_label_id" appended to the \code{column_name} will be created.}
}
\description{
This function associates a label (referenced by a label_id) to an image (referenced by image_id).
The a merge between two dataframes will be made using
A POST request is made to api/v1/images/:id/labels
See https://biigle.de/doc/api/index.html#api-Images-StoreImageLabels
}
